chore: created separated package for PrivSigner structure
