+++ app/signer/private.go
@@ -1,0 +1,33 @@
+package signer
+
+import (
+	"path/filepath"
+
+	cmtconfig "github.com/cometbft/cometbft/config"
+	cmtos "github.com/cometbft/cometbft/libs/os"
+
+	"github.com/babylonlabs-io/babylon/privval"
+)
+
+type PrivSigner struct {
+	WrappedPV *privval.WrappedFilePV
+}
+
+func InitPrivSigner(nodeDir string) (*PrivSigner, error) {
+	nodeCfg := cmtconfig.DefaultConfig()
+	pvKeyFile := filepath.Join(nodeDir, nodeCfg.PrivValidatorKeyFile())
+	err := cmtos.EnsureDir(filepath.Dir(pvKeyFile), 0777)
+	if err != nil {
+		return nil, err
+	}
+	pvStateFile := filepath.Join(nodeDir, nodeCfg.PrivValidatorStateFile())
+	err = cmtos.EnsureDir(filepath.Dir(pvStateFile), 0777)
+	if err != nil {
+		return nil, err
+	}
+	wrappedPV := privval.LoadOrGenWrappedFilePV(pvKeyFile, pvStateFile)
+
+	return &PrivSigner{
+		WrappedPV: wrappedPV,
+	}, nil
+}
